requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.3
python-dotenv==0.15.0
Werkzeug==1.0.1
mysql-connector-python
mysql
